+++ test/unit/config-loader.test.ts
@@ -1,3 +1,4 @@
 /** Entry point for split config-loader.test.ts. */
 import "./config-loader-01.test.ts";
 import "./config-loader-02.test.ts";
+import "./config-loader-environment.test.ts";